cork-on-phone: Only cork (and subsequently uncork) streams that are not already corked.
Although by "cork" I really mean "cork+mute" as that's what the module
does.

If e.g. Rhythmbox is paused when a phone call comes in, the current
stream state will be corked and thus we should not track it for future
uncorking when the phone call ends.

Likewise if the stream is just muted (manually) we will not take any
action either when the phone stream is seen first, nor when it
disappears.

Also add some additional debug messages.