pa_poll(): Simplify detection of invalid fds in select() emulation mode
For systems which have a fcntl() implementation, we can simplify the
code which determines whether a file selector is valid in pa_poll().

The old code, which is harder to read and more expensive, stays around
for all platforms we need to emulate poll() for using select(), and
which don't provide fcntl(). IOW, for Windows.

On Mac OS X, however, the detection for bad fds via more select() calls
doesn't work, resulting in hung main loops, so the patch fixes a real
bug there.